feat: large document support (#27)
Chunks the server update state into 10_000 character chunks.

This works around the 128 KB value size limit for Cloudflare durable storage.

Prior to this change 128 KB was the maximum document size.

Tested with 1, 5 and 10 MB documents.

With this chunking, many edits (adding new lines, editing in the middle of the document, etc), result in all chunks being poked from the server to the clients. So large documents are inefficient but functioning.

We should try to replace this basic chunking approach with a [content-defined chunking approach](https://joshleeb.com/posts/content-defined-chunking.html).