perf: Don't send mutations for updates originating from this provider. (#35)
If we don't do this then when a provider receives an update from the server it does not yet have and applies it to the document it resends this update in a mutation to the server which then pokes it out to all other clients.  The loop stops there as the clients all already have the update and so don't emit it again.  Echoing the update back and sending no-op pokes is inefficient, so don't do it.

See https://github.com/yjs/yjs/issues/321#issuecomment-890534515