perf: change mutation updates to not be based on server vector (#37)
The updates in mutations are now just the update message fired by the ydoc.  It is not necessary for these to be based on a vector of a previous server state.  As long as all update messages are propagated to all clients, all clients will eventually converge to the same state.  This is ensured by the reflect mutation/push system ensuring all updates flow from a source ydoc to the server, where they are merged into the server update which is then poked to all clients.

Since the updates in mutations are no longer cumulative each client update is written at its own key (using a uuid).   These are not written by the server handling of the mutation, and thus are effectively deleted once they are merged into the server update.

Tested loading of the doc from offline using `kvStore: 'idb'` and also offline persist/refresh.