perf: shorten hashes from 64 chars to 12 chars (#41)
Hashes are currently hex encoded sha-256 digests, resulting in a hash length of 64 chars.  This is overkill and creates significant overhead in the chunk keys and in the meta array of all chunk hashes.  

Reduce the hash length from 64 chars to 12 chars as 81.25% reduction in size.  This is done by truncating the sha-256 digest to 9 bytes (72 bits) and base64 encoding it.  

This still gives us plenty of collision resistance for the range of expected document sizes.
If we assume a max document size of 100MB, the probability of having a
hash collision in a 100MB document is roughly 1 in a trillion.
 
This is based on the approximation function p(n) = n^2 / (2H) from
https://en.wikipedia.org/wiki/Birthday_attack#Simple_approximation.
Where p(n) is probability of collision, n is number of hashes, and H is number of 
possible hash outputs.  For 100MB document we have
 n = 100,000 (100,000 1KB chunks) and H = 2^(8*9),
 
 p(100,000) = (100,000^2 / (2*2^72)) = 1.0587912e-12